requirements.txt
torch
torchvision
numpy
flask
Pillow
matplotlib
